+++ src/main/kotlin/net/trueog/diamondbankog/ErrorHandler.kt
@@ -10,7 +10,7 @@ internal object ErrorHandler {
     /** Handles the error by throwing, disables the economy unless you specify it shouldn't */
     fun handleError(
         uuid: UUID,
-        expectedMutatedShards: Int,
+        expectedMutatedShards: Long,
         playerShards: PlayerShards?,
         otherUuid: UUID? = null,
         dontDisableEconomy: Boolean = false,
@@ -25,15 +25,15 @@ internal object ErrorHandler {
                 if (otherUuid != null) "Other Player UUID: $otherUuid" else ""
             }Expected Mutated Shards = $expectedMutatedShards${
                 if (playerShards != null) {
-                    if (playerShards.bank != -1) "Player Bank Balance: ${playerShards.bank}" else ""
+                    if (playerShards.bank != -1L) "Player Bank Balance: ${playerShards.bank}" else ""
                 } else ""
             }${
                 if (playerShards != null) {
-                    if (playerShards.inventory != -1) "Player Inventory Balance: ${playerShards.inventory}" else ""
+                    if (playerShards.inventory != -1L) "Player Inventory Balance: ${playerShards.inventory}" else ""
                 } else ""
             }${
                 if (playerShards != null) {
-                    if (playerShards.enderChest != -1) "Player Ender Chest Balance: ${playerShards.enderChest}" else ""
+                    if (playerShards.enderChest != -1L) "Player Ender Chest Balance: ${playerShards.enderChest}" else ""
                 } else ""
             }
         """
